[knex] Transaction should mirror Knex (see http://knexjs.org/#Transactions)
The documentation states:
```
Transactions are handled by passing a handler function into
knex.transaction. The handler function accepts a single argument,
an object which may be used in two ways:

As the "promise aware" knex connection
As an object passed into a query with and eventually call commit or
rollback.
```

Meaning that the Transaction type should mirror the Knex object in
functionality, including .schema, .raw, and ('tablename')